Trim down our docker image by about 400MB
Every RUN command in a Dockerfile creates a new 'layer', and each of
those layers persists, even if you delete files later on. So when it
comes to things like NPM installs, package installations from APT, or
any other file-generative operations, it is a good idea to clean up
after yourself within the same RUN directive.

This overhaul of the Dockerfile does just that, and trims us down to a
single RUN command that cleans up after itself as best as I'm aware we                  can. We're even deleting `node_modules`. Which is fun.

I don't know that we can trim it much more without creating our own
squashed fc-base image. For now though this should help.